[FEAT/#22] 멀티모듈 의존성 추가
