+++ app/src/main/java/com/sopt/now/compose/feature/User.kt
@@ -1,12 +1,12 @@
 package com.sopt.now.compose.feature
 
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class User(
-    val id: String,
-    val pw: String,
-    val nickname: String,
-    val phone: String
-) : Parcelable
+//import android.os.Parcelable
+//import kotlinx.parcelize.Parcelize
+//
+//@Parcelize
+//data class User(
+//    val id: String,
+//    val pw: String,
+//    val nickname: String,
+//    val phone: String
+//) : Parcelable
