+++ data/build.gradle
@@ -30,6 +30,7 @@ android {
 }
 
 dependencies {
+    implementation(project(":domain"))
 
     implementation 'androidx.core:core-ktx:1.13.1'
     implementation 'androidx.appcompat:appcompat:1.6.1'
