+++ feature/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
+    id 'kotlin-parcelize'
 }
 
 android {
@@ -30,6 +31,8 @@ android {
 }
 
 dependencies {
+    implementation(project(":domain"))
+    implementation(project(":core-ui"))
 
     implementation 'androidx.core:core-ktx:1.13.1'
     implementation 'androidx.appcompat:appcompat:1.6.1'
