+++ feature/src/main/java/com/sopt/feature/User.kt
@@ -1,0 +1,12 @@
+package com.sopt.feature
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class User(
+    val id: String,
+    val pw: String,
+    val nickname: String,
+    val phone: String
+) : Parcelable
